Centralize model consistency checker test assertion
